+++ src/javascript/utils/running.config.js
@@ -3,7 +3,8 @@ import {puffin} from '@mkenzo_8/puffin'
 let Config = {
     focusedPanel:null,
     focusedTab:null,
-    focusedEditor:null
+    focusedEditor:null,
+    openedFolders:[]
 }
 
 const RunningConfig = new puffin.state(Config)